Parse lcov with branch summaries but no branches.
Parse an `lcov.info` file that contains `BRF` and `BRH` records but no
`BRDA` records. To support this, each `item` object is created with a
`branches` property. The `branches` property is not only created if
needed. To determine if there are no branches in an `item` object,
instead of testing for the absence of the `branches` property, the user
should instead to the `found` property of the `branches` property.

Closes #3.